Enhance folders and code challenge descriptions
