+++ src/pascals-triangle/packagepascalstriangle/Main.java
@@ -1,4 +1,4 @@
-package packagePascalsTriangle;
+package packagepascalstriangle;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -7,26 +7,6 @@ import java.util.stream.IntStream;
 
 public class Main {
 
-//	Challenge:
-//	
-//	Given an integer numRows, return the first numRows of Pascal's triangle.
-//	In Pascal's triangle, each number is the sum of the two numbers directly above it.
-//	
-//	Example 1:
-//
-//	Input: numRows = 5
-//	Output: [[1],[1,1],[1,2,1],[1,3,3,1],[1,4,6,4,1]]
-//			
-//	Example 2:
-//
-//	Input: numRows = 1
-//	Output: [[1]]
-//				 
-//
-//	Constraints:
-//
-//	1 <= numRows <= 30
-      
   public static void main(String[] args) {
 
     int numRows = 5;
